@@ -1,8 +1,8 @@
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
   theme: {
-    colors: { 'orange': '#FF8A00' },
-    fontFamily: {
-      'pfont': 'Barlow Semi Condensed', //fonte dos paragrafos <p>
-    },
+    extend: {},
   },
+  plugins: [],
 };
