@@ -1,9 +1,9 @@
 import { useState } from 'react';
-// import reactLogo from './assets/react.svg';
+import reactLogo from './assets/react.svg';
 import viteLogo from '/vite.svg';
-// import './App.css';
+import './App.css';
 
-function Home() {
+function App() {
   const [count, setCount] = useState(0);
 
   return (
@@ -13,7 +13,7 @@ function Home() {
           <img src={viteLogo} className="logo" alt="Vite logo" />
         </a>
         <a href="https://react.dev" target="_blank">
-          {/* <img src={reactLogo} className="logo react" alt="React logo" /> */}
+          <img src={reactLogo} className="logo react" alt="React logo" />
         </a>
       </div>
       <h1>Vite + React</h1>
@@ -25,11 +25,9 @@ function Home() {
           Edit <code>src/App.tsx</code> and save to test HMR
         </p>
       </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
+      <p className="">Click on the Vite and React logos to learn more</p>
     </>
   );
 }
 
-export default Home;
+export default App;
